+++ src/polynx/__init__.py
@@ -1,9 +1,13 @@
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 from . import core as _core
-from .utils import plx_frame_patch, plx_expr_patch, plx_rolling_prod, DataFrame, LazyFrame, Series, Expr, plx_merge as merge
+from .utils import plx_frame_patch, plx_expr_patch, DataFrame, LazyFrame, Series, Expr, plx_merge as merge
 from . import io as _io
 
 __all__ = ["DataFrame", "LazyFrame", "Series", "Expr"]
+
+for name in ['rolling_prod']:
+    plx_expr_patch(name, getattr(_core, f"plx_{name}"))
+
 
 for name in [
     'eval','assign','dd', "dsort", "asort", "unstack", 
@@ -13,7 +17,6 @@ for name in [
 ]:
     plx_frame_patch(name, getattr(_core, f"plx_{name}"))
 
-plx_expr_patch('rolling_prod', plx_rolling_prod)
 
 for name in _io._wrapped_functions:
     globals()[name] = getattr(_io, name)