+++ src/polynx/expr.py
@@ -4,19 +4,20 @@ from .wrapper import wrap, unwrap
 class Expr:
     def __init__(self, *args, **kwargs):
         self._pl = pl.Expr(*args, **kwargs)
+        self._is_wrapped = True
 
     def __getattr__(self, name):
-        attr = getattr(self._pl, name)
-
-        if callable(attr):
-            def wrapped(*args, **kwargs):
-                args = [unwrap(a) for a in args]
-                kwargs = {k: unwrap(v) for k, v in kwargs.items()}
-                result = attr(*args, **kwargs)
-                return wrap(result)
-            return wrapped
-
-        return attr  
+        try:
+            return super().__getattribute__(name)
+        except AttributeError:
+            _pl = super().__getattribute__("_pl")
+            attr = getattr(_pl, name)
+            if callable(attr):
+                def wrapped(*args, **kwargs):
+                    from .wrapper import wrap, unwrap
+                    return wrap(attr(*[unwrap(a) for a in args], **{k: unwrap(v) for k, v in kwargs.items()}))
+                return wrapped
+            return attr  
 
     def __repr__(self):
         return f"PolynxExpr:\n{repr(self._pl)}"